Plumb outputs to console.
