+++ packages/shared-ui/src/state/work-item.ts
@@ -10,11 +10,10 @@ import {
   LLMContent,
   NodeEndResponse,
   NodeTypeIdentifier,
-  OutputValues,
 } from "@breadboard-ai/types";
 import { SignalMap } from "signal-utils/map";
 import { InputResponse, OutputResponse, Schema } from "@google-labs/breadboard";
-import { idFromPath } from "./common";
+import { idFromPath, toLLMContentArray } from "./common";
 
 export { ReactiveWorkItem };
 
@@ -83,67 +82,5 @@ class ReactiveWorkItem implements WorkItem {
     for (const [name, product] of Object.entries(products)) {
       item.product.set(name, product);
     }
-  }
-}
-
-function toLLMContentArray(
-  schema: Schema,
-  values: OutputValues
-): { products: Record<string, LLMContent>; chat: boolean } {
-  let chat = false;
-  if (!schema.properties) {
-    // No schema, so let's just stringify and stuff outputs into json part.
-    const products = Object.fromEntries(
-      Object.entries(values).map(([name, value]) => {
-        return [name, asJson(value)];
-      })
-    );
-    return { products, chat };
-  }
-
-  const products: Record<string, LLMContent> = {};
-  for (const [name, propertySchema] of Object.entries(schema.properties)) {
-    const value = values[name];
-    if (!value) {
-      console.warn(
-        `Schema specifies property "${name}", but it wasn't supplied`
-      );
-      continue;
-    }
-    if (propertySchema.behavior?.includes("hint-chat-mode")) {
-      chat = true;
-    }
-    if (propertySchema.type === "array") {
-      const items = propertySchema.items as Schema;
-      if (items.behavior?.includes("llm-content")) {
-        // This is an LLMContent array. By convention, we only take the first
-        // item.
-        if (Array.isArray(value) && value.length > 0) {
-          products[name] = value.at(0) as LLMContent;
-          continue;
-        }
-      }
-    } else if (
-      propertySchema.type === "object" &&
-      propertySchema.behavior?.includes("llm-content")
-    ) {
-      // This is an LLMContent.
-      products[name] = value as LLMContent;
-      continue;
-    } else if (
-      propertySchema.type === "string" ||
-      propertySchema.type === "number" ||
-      propertySchema.type === "boolean"
-    ) {
-      products[name] = { parts: [{ text: `${value}` }] };
-      continue;
-    }
-    // Everything else, let's stringify and stuff outputs as json part.
-    products[name] = asJson(value);
-  }
-  return { products, chat };
-
-  function asJson(value: unknown): LLMContent {
-    return { parts: [{ json: JSON.stringify(value, null, 2) }] };
   }
 }
